requirements.txt
numpy==1.19.4
librosa==0.8.0
matplotlib==3.1.1
torch==1.6.0
torchvision==0.7.0
Pillow==8.3.2
pydub==0.24.1
spotipy==2.19.0
pandas==1.1.5
requests==2.26.0
tqdm==4.62.3
scikit-learn==1.0